build(lib): change index to lib index
