--- src/index.js
@@ -1,16 +1,0 @@
-const components = {};
-
-function importAll(r) {
-    r.keys().forEach(key => {
-        const name = key.replace(/(^\.\/)|(\/index\.jsx$)/g, '');
-        const all = r(key);
-        const component = all.default;
-        components[name] = Object.assign(component, all);
-    });
-}
-
-importAll(require.context('./components/', true, /index\.jsx$/));
-
-delete components.List;
-
-module.exports = components;
